MedDataOS/requirements.txt
# Core essentials
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Model persistence
joblib>=1.3.0

# Basic visualizations
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment management
python-dotenv>=1.0.0 

# Google Generative AI
google-genai>=1.0.0

# LangGraph and dependencies
langgraph>=0.0.19
langchain>=0.1.0
pydantic>=2.0.0

# Web UI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6  # For file uploads in FastAPI

# Excel file support
openpyxl>=3.1.0  # For pandas to read .xlsx files 